feat: rename PraxisInstance file
