+++ .env.example
@@ -1,3 +1,4 @@
+NODE_ENV=development
 BOT_TOKEN=YOUR_BOT_TOKEN
 PORT=3200
